Move algorithms under hca pkg
